# internal/server/routes/routes.go
// Package routes defines API route handlers for Lighthouse.
package routes

import (
	"time"

	"github.com/gin-gonic/gin"
	"github.com/myxxhui/lighthouse-src/internal/server/dto"
)

// RegisterCostRoutes registers all cost-related routes.
func RegisterCostRoutes(group *gin.RouterGroup) {
	// Global cost overview
	group.GET("/global", getGlobalCost)
	// Namespace cost
	group.GET("/namespace/:namespace", getNamespaceCost)
	// Drilldown
	group.GET("/drilldown/:level/:identifier", getDrilldownCost)
}

// RegisterSLORoutes registers all SLO-related routes.
func RegisterSLORoutes(group *gin.RouterGroup) {
	group.GET("/health", getSLOHealth)
	group.GET("/history", getSLOHistory)
	group.GET("/burnrate", getSLOBurnRate)
}

// RegisterROIRoutes registers all ROI-related routes.
func RegisterROIRoutes(group *gin.RouterGroup) {
	group.GET("/dashboard", getROIDashboard)
	group.GET("/details", getROIDetails)
	group.GET("/comparison", getROIComparison)
}

// =============================================
// Cost Route Handlers (to be implemented in cost.go)
// =============================================

func getGlobalCost(c *gin.Context) {
	// TODO: Implement business logic integration
	c.JSON(200, dto.GlobalCostResponse{
		TotalCost: 10000.0,
		Namespaces: []dto.NamespaceCostSummary{
			{Name: "default", Cost: 5000.0, PodCount: 10, NodeCount: 2},
			{Name: "kube-system", Cost: 3000.0, PodCount: 5, NodeCount: 1},
			{Name: "monitoring", Cost: 2000.0, PodCount: 3, NodeCount: 1},
		},
		Timestamp: time.Now().UTC(),
	})
}

func getNamespaceCost(c *gin.Context) {
	// TODO: Implement
	c.JSON(200, dto.NamespaceCostResponse{
		Namespace: c.Param("namespace"),
		Cost: dto.CostBreakdown{
			Total:      5000.0,
			CPU:        3000.0,
			Memory:     2000.0,
			Billable:   4000.0,
			Usage:      800.0,
			Waste:      200.0,
			Efficiency: 0.85,
		},
		Timestamp: time.Now().UTC(),
	})
}

func getDrilldownCost(c *gin.Context) {
	// TODO: Implement
	c.JSON(200, dto.DrilldownResponse{
		Level:      c.Param("level"),
		Identifier: c.Param("identifier"),
		Cost: dto.CostBreakdown{
			Total: 2500.0,
		},
		Timestamp: time.Now().UTC(),
	})
}

// =============================================
// SLO Route Handlers (to be implemented in slo.go)
// =============================================

func getSLOHealth(c *gin.Context) {
	// TODO: Implement
	c.JSON(200, dto.SLOHealthResponse{
		Status:    "healthy",
		Timestamp: time.Now().UTC(),
	})
}

func getSLOHistory(c *gin.Context) {
	// TODO: Implement
	c.JSON(200, dto.SLOHistoryResponse{
		Service:   c.Query("service"),
		Timestamp: time.Now().UTC(),
	})
}

func getSLOBurnRate(c *gin.Context) {
	// TODO: Implement
	c.JSON(200, dto.SLOBurnRateResponse{
		Service:   c.Query("service"),
		Timestamp: time.Now().UTC(),
	})
}

// =============================================
// ROI Route Handlers (to be implemented in roi.go)
// =============================================

func getROIDashboard(c *gin.Context) {
	// TODO: Implement
	c.JSON(200, dto.ROIDashboardResponse{
		Summary: dto.ROISummary{
			ROIPercentage: 45.2,
			TotalSavings:  125000.0,
			Status:        "good",
			Trend:         "improving",
		},
		Timestamp: time.Now().UTC(),
	})
}

func getROIDetails(c *gin.Context) {
	// TODO: Implement
	c.JSON(200, dto.ROIDetailsResponse{
		Category:  c.Query("category"),
		Timestamp: time.Now().UTC(),
	})
}

func getROIComparison(c *gin.Context) {
	// TODO: Implement
	c.JSON(200, dto.ROIComparisonResponse{
		Timestamp: time.Now().UTC(),
	})
}